rasa/actions/requirements.txt
rasa==3.6.0
rasa-sdk==3.6.0
python-dotenv==1.0.0

